Extracted inputs for proposal form to their own components
