+++ src/components/pageBody/proposal/ProposalSubmit.tsx
@@ -1,280 +1,41 @@
-import {
-  Button,
-  Flex,
-  Input,
-  InputGroup,
-  InputLeftAddon,
-  Heading,
-  NumberInput,
-  NumberInputField,
-  NumberInputStepper,
-  NumberIncrementStepper,
-  NumberDecrementStepper,
-  RadioGroup,
-  Stack,
-  Radio,
-  Slider,
-  SliderTrack,
-  SliderFilledTrack,
-  SliderThumb
-} from '@chakra-ui/react';
+import { Button, Flex, Heading } from '@chakra-ui/react';
 import useHandleClicks from 'hooks/useHandleClicks';
-import useHandleInputs from 'hooks/useHandleInputs';
-import { useRecoilState, useRecoilValue } from 'recoil';
-import {
-  inputDevelopmentCost,
-  inputExpiration,
-  inputGitHub,
-  inputNameProposal,
-  inputPaymentFrequency,
-  inputPaymentSchedule,
-  inputQuorumPercent,
-  inputSelfVoucedInformation,
-  inputStatementOfWork,
-  inputTvlLimit,
-  inputUrlProposal,
-  inputInitialFunding
-} from 'recoil/input/atoms';
+import InputDevelopmentCost from './inputs/InputDevelopmentCost';
+import InputExpiration from './inputs/InputExpiration';
+import InputGithub from './inputs/InputGithub';
+import InputInitialFunding from './inputs/InputInitialFunding';
+import InputPaymentFrequency from './inputs/InputPaymentFrequency';
+import InputPaymentSchedule from './inputs/InputPaymentSchedule';
+import InputProposalName from './inputs/InputProposalName';
+import InputProposalUrl from './inputs/InputProposalUrl';
+import InputQuorumPercentage from './inputs/InputQuorumPercentage';
+import InputSelfVouchedInformation from './inputs/InputSelfVouchedInformation';
+import InputStatementOfWork from './inputs/InputStatementOfWork';
+import InputTvlLimit from './inputs/InputTvlLimit';
 
 function ProposalSubmit() {
-  const [paymentSchedule, setPaymentSchedule] =
-    useRecoilState(inputPaymentSchedule);
-
-  const nameProposal = useRecoilValue(inputNameProposal);
-  const urlProposal = useRecoilValue(inputUrlProposal);
-  const tvlLimit = useRecoilValue(inputTvlLimit);
-  const developmentCost = useRecoilValue(inputDevelopmentCost);
-  const statementOfWork = useRecoilValue(inputStatementOfWork);
-  const github = useRecoilValue(inputGitHub);
-  const quorumPercent = useRecoilValue(inputQuorumPercent);
-  const selfVouchedInformation = useRecoilValue(inputSelfVoucedInformation);
-  const expiration = useRecoilValue(inputExpiration);
-  const paymentFrequency = useRecoilValue(inputPaymentFrequency);
-  const initialFunding = useRecoilValue(inputInitialFunding);
-
   const { handleClickCreateProposal } = useHandleClicks();
-
-  const {
-    handleInputNameProposal,
-    handleInputUrlProposal,
-    handleInputTvlLimit,
-    handleInputDevelopmentCost,
-    handleInputInitialFunding,
-    handleInputStatementOfWork,
-    handleInputPaymentFrequency,
-    handleInputGithub,
-    handleInputQuorumPercent,
-    handleInputSelfVouchedInformation,
-    handleInputExpiration
-  } = useHandleInputs();
 
   return (
     <Flex direction="column" alignItems="center" minWidth="50%" mt={5}>
       <Heading as="h2" size="lg">
         Submit a Proposal
       </Heading>
-      <Input
-        mt={5}
-        value={nameProposal}
-        onChange={handleInputNameProposal}
-        placeholder="Enter Proposal Name"
-      />
+      <InputProposalName />
+      <InputProposalUrl />
+      <InputStatementOfWork />
+      <InputGithub />
+      <InputSelfVouchedInformation />
 
-      <InputGroup>
-        <InputLeftAddon children="https://" />
-        <Input
-          value={urlProposal}
-          onChange={handleInputUrlProposal}
-          placeholder="Enter Proposal Url"
-        />
-      </InputGroup>
-      <InputGroup>
-        <InputLeftAddon children="https://" />
-        <Input
-          value={statementOfWork}
-          onChange={handleInputStatementOfWork}
-          placeholder="Enter URL to statement of work"
-        />
-      </InputGroup>
-      <InputGroup>
-        <InputLeftAddon children="https://" />
-        <Input
-          value={github}
-          onChange={handleInputGithub}
-          placeholder="Enter URL to Github"
-        />
-      </InputGroup>
-      <InputGroup>
-        <InputLeftAddon children="https://" />
-        <Input
-          value={selfVouchedInformation}
-          onChange={handleInputSelfVouchedInformation}
-          placeholder="Enter URL to Self Vouched Information"
-        />
-      </InputGroup>
-      <Flex alignItems="center" gap={5} mt={5}>
-        <Heading as="h3" size="md">
-          Maximum Vault TVL:
-        </Heading>
-        <NumberInput
-          maxW="140px"
-          mr="2rem"
-          defaultValue={1000000}
-          step={1000}
-          min={1000000}
-          value={tvlLimit}
-          onChange={handleInputTvlLimit}
-        >
-          <NumberInputField />
-          <NumberInputStepper>
-            <NumberIncrementStepper />
-            <NumberDecrementStepper />
-          </NumberInputStepper>
-        </NumberInput>
-      </Flex>
+      <InputTvlLimit />
+      <InputDevelopmentCost />
+      <InputInitialFunding />
+      <InputExpiration />
 
-      <Flex alignItems="center" gap={5}>
-        <Heading as="h3" size="md">
-          Development Cost in YFD:
-        </Heading>
-        <NumberInput
-          maxW="140px"
-          mr="2rem"
-          defaultValue={5000}
-          step={1}
-          min={0}
-          // max={1000000}
-          value={developmentCost}
-          onChange={handleInputDevelopmentCost}
-        >
-          <NumberInputField />
-          <NumberInputStepper>
-            <NumberIncrementStepper />
-            <NumberDecrementStepper />
-          </NumberInputStepper>
-        </NumberInput>
-      </Flex>
-      <Flex alignItems="center" gap={5}>
-        <Heading as="h3" size="md">
-          Initial Funding in YFD:
-        </Heading>
-        <NumberInput
-          maxW="140px"
-          mr="2rem"
-          defaultValue={0}
-          step={1}
-          min={0}
-          max={developmentCost}
-          value={initialFunding}
-          onChange={handleInputInitialFunding}
-        >
-          <NumberInputField />
-          <NumberInputStepper>
-            <NumberIncrementStepper />
-            <NumberDecrementStepper />
-          </NumberInputStepper>
-        </NumberInput>
-      </Flex>
+      <InputPaymentSchedule />
+      <InputPaymentFrequency />
 
-      <Flex alignItems="center" gap={5}>
-        <Heading as="h3" size="md">
-          Expiration in Blocks:
-        </Heading>
-        <NumberInput
-          maxW="140px"
-          mr="2rem"
-          defaultValue={43200}
-          step={1}
-          min={43200}
-          value={expiration}
-          onChange={handleInputExpiration}
-        >
-          <NumberInputField />
-          <NumberInputStepper>
-            <NumberIncrementStepper />
-            <NumberDecrementStepper />
-          </NumberInputStepper>
-        </NumberInput>
-      </Flex>
-
-      <RadioGroup
-        as={Flex}
-        my={5}
-        direction="column"
-        alignItems="center"
-        onChange={setPaymentSchedule}
-        value={paymentSchedule}
-      >
-        <Heading as="h3" size="md">
-          Select Payment Schedule
-        </Heading>
-        <Stack direction="row">
-          <Radio value="0">All at Beginning</Radio>
-          <Radio value="1">All at Completion</Radio>
-          <Radio value="2">Half at Frequency, Half at Completion</Radio>
-          <Radio value="3">Thirds</Radio>
-        </Stack>
-      </RadioGroup>
-
-      <Flex alignItems="center" gap={5}>
-        <Heading as="h3" size="md">
-          Payment Frequency in Blocks:
-        </Heading>
-        <NumberInput
-          maxW="140px"
-          mr="2rem"
-          defaultValue={1000}
-          step={1}
-          min={1000}
-          value={paymentFrequency}
-          onChange={handleInputPaymentFrequency}
-        >
-          <NumberInputField />
-          <NumberInputStepper>
-            <NumberIncrementStepper />
-            <NumberDecrementStepper />
-          </NumberInputStepper>
-        </NumberInput>
-      </Flex>
-
-      <Flex width="100%" alignItems="center" gap={5}>
-        <Heading as="h3" size="md">
-          Quorum Percentage:
-        </Heading>
-        <NumberInput
-          maxW="140px"
-          mr="2rem"
-          my={5}
-          defaultValue={25}
-          step={1}
-          min={5}
-          max={95}
-          value={quorumPercent}
-          onChange={handleInputQuorumPercent}
-        >
-          <NumberInputField />
-          <NumberInputStepper>
-            <NumberIncrementStepper />
-            <NumberDecrementStepper />
-          </NumberInputStepper>
-        </NumberInput>
-        <Slider
-          flex="1"
-          focusThumbOnChange={false}
-          defaultValue={25}
-          step={1}
-          min={5}
-          max={95}
-          value={quorumPercent}
-          onChange={handleInputQuorumPercent}
-        >
-          <SliderTrack>
-            <SliderFilledTrack />
-          </SliderTrack>
-          <SliderThumb fontSize="sm" boxSize="12px" children={quorumPercent} />
-        </Slider>
-      </Flex>
-
+      <InputQuorumPercentage />
       <Button
         my={5}
         onClick={() => {
